packages/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codeverify-core"
version = "0.2.0"
description = "CodeVerify Core Utilities"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codeverify_core"]
